refactor(ai): split oversized expand planner test files (#3941)
Decompose expand_phase_planner_test.dart and expand_phase_planner_declare_war_test.dart
into concern-focused suites (peace, predicates, declare-war core, H2 cooldown, S7-D
refutation) with 1:1 test preservation.

Refs #3941
