+++ packages/colonizethis_ai/test/planning/expand_phase_planner_declare_war_test.dart
@@ -605,324 +605,5 @@ void main() {
       );
       expect(planExpandDeclareWar(game: game, snapshot: snapshot), isNull);
     });
-
-    test('Refs #2847 § H2: peer-war peace cooldown active -> null '
-        '(arm 3 suppressed during cooldown window)', () {
-      // Refs #2847 § H2 positive case.
-      //
-      // Seed-42 post-H4-a refresh shape: gp1 (active player, 8 OW) and
-      // peer gp2 (8 OW) just made peace last turn (the H4-a carve-out
-      // in `planExpandPeace` fired and the mutual offer completed).
-      // The next turn, every other arm-3 gate still passes —
-      // treasury (9999 >= cheapest), regiment parity (5 == 5),
-      // mutual-plateau (8 vs 8 below quota), gp2 sole adjacent OW
-      // owner, gp2 not in atWarWith. Without the H2 cooldown the
-      // planner would re-declare on gp2 the very next turn and the
-      // war would re-open. The cooldown gate must short-circuit arm 3
-      // and return null while a peace event between {_gp1, _gp2} sits
-      // within the last `kExpandPeerWarPeaceCooldownTurns` turns.
-      final owProvinces = <Province>[
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp1_$i', regionId: 'oldWorld', ownerId: _gp1),
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp2_$i', regionId: 'oldWorld', ownerId: _gp2),
-      ];
-      final game = buildExpandGame(
-        gameIdLabel: 'expand-phase-planner-declare-war',
-        turnNumber: 50,
-        oldWorldProvinces: owProvinces,
-        armies: [
-          homeArmyWithRegimentsAtCapital(_gp1, 5),
-          homeArmyWithRegimentsAtCapital(_gp2, 5),
-        ],
-        diplomaticHistoryEvents: const [
-          DiplomaticEvent(
-            turn: 49,
-            intraTurnIndex: 0,
-            type: DiplomaticEventType.peace,
-            participants: {_gp1, _gp2},
-            fromFactionId: _gp1,
-            toFactionId: _gp2,
-          ),
-        ],
-      );
-      final snapshot = buildExpandSnapshot(
-        atWarWith: const [],
-        invadableOw: const ['oldWorld|gp2_0'],
-        adjacentOwners: const [_gp2],
-        oldWorldProvincesOwned: 8,
-      );
-      expect(
-        planExpandDeclareWar(game: game, snapshot: snapshot),
-        isNull,
-        reason:
-            'Peace event between {gp1, gp2} on turn 49 is 1 turn old '
-            'on turn 50; with kExpandPeerWarPeaceCooldownTurns = 4 '
-            'the cooldown is active -> arm 3 must return null even '
-            'though every other gate passes (treasury, regiments, '
-            'mutual-plateau, sole GP blocker). Refs #2847 § H2.',
-      );
-    });
-
-    test('Refs #2847 § H2: peer-war peace cooldown lapsed -> blocker GP '
-        '(arm 3 fires once cooldown window expires)', () {
-      // Refs #2847 § H2 boundary case (cooldown lapsed).
-      //
-      // Same fixture as the positive case but the peace event is
-      // exactly `kExpandPeerWarPeaceCooldownTurns` turns old (turn 46
-      // peace, current turn 50). The strict `<` boundary inside
-      // `expandRecentlyPeacedWithGreatPower` means the cooldown is
-      // **not** active and arm 3 must fire normally, returning gp2.
-      final owProvinces = <Province>[
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp1_$i', regionId: 'oldWorld', ownerId: _gp1),
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp2_$i', regionId: 'oldWorld', ownerId: _gp2),
-      ];
-      final game = buildExpandGame(
-        gameIdLabel: 'expand-phase-planner-declare-war',
-        turnNumber: 50,
-        oldWorldProvinces: owProvinces,
-        armies: [
-          homeArmyWithRegimentsAtCapital(_gp1, 5),
-          homeArmyWithRegimentsAtCapital(_gp2, 5),
-        ],
-        diplomaticHistoryEvents: const [
-          DiplomaticEvent(
-            turn: 46,
-            intraTurnIndex: 0,
-            type: DiplomaticEventType.peace,
-            participants: {_gp1, _gp2},
-            fromFactionId: _gp1,
-            toFactionId: _gp2,
-          ),
-        ],
-      );
-      final snapshot = buildExpandSnapshot(
-        atWarWith: const [],
-        invadableOw: const ['oldWorld|gp2_0'],
-        adjacentOwners: const [_gp2],
-        oldWorldProvincesOwned: 8,
-      );
-      expect(
-        planExpandDeclareWar(game: game, snapshot: snapshot),
-        _gp2,
-        reason:
-            'Peace event on turn 46 is exactly 4 turns old on turn 50; '
-            'the cooldown predicate uses `currentTurn - event.turn < '
-            'kExpandPeerWarPeaceCooldownTurns` so 4 < 4 is false -> '
-            'cooldown lapses and arm 3 returns gp2. Refs #2847 § H2 '
-            'boundary.',
-      );
-    });
-
-    test('Refs #2847 § H2: cooldown applies symmetrically (peace event '
-        'recorded with the peer as fromFactionId still suppresses)', () {
-      // Refs #2847 § H2 symmetry pin.
-      //
-      // The peace event was finalized when gp2 (the peer) was the
-      // mutual-offer second leg, so `fromFactionId: gp2` and
-      // `toFactionId: gp1`. The H2 cooldown looks at `participants`
-      // (a Set) only, so the active player gp1 must still see the
-      // cooldown as active on turn 50. Guards against a regression
-      // where the helper accidentally requires the active player to
-      // be `event.fromFactionId`.
-      final owProvinces = <Province>[
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp1_$i', regionId: 'oldWorld', ownerId: _gp1),
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp2_$i', regionId: 'oldWorld', ownerId: _gp2),
-      ];
-      final game = buildExpandGame(
-        gameIdLabel: 'expand-phase-planner-declare-war',
-        turnNumber: 50,
-        oldWorldProvinces: owProvinces,
-        armies: [
-          homeArmyWithRegimentsAtCapital(_gp1, 5),
-          homeArmyWithRegimentsAtCapital(_gp2, 5),
-        ],
-        diplomaticHistoryEvents: const [
-          DiplomaticEvent(
-            turn: 49,
-            intraTurnIndex: 0,
-            type: DiplomaticEventType.peace,
-            participants: {_gp1, _gp2},
-            fromFactionId: _gp2,
-            toFactionId: _gp1,
-          ),
-        ],
-      );
-      final snapshot = buildExpandSnapshot(
-        atWarWith: const [],
-        invadableOw: const ['oldWorld|gp2_0'],
-        adjacentOwners: const [_gp2],
-        oldWorldProvincesOwned: 8,
-      );
-      expect(
-        planExpandDeclareWar(game: game, snapshot: snapshot),
-        isNull,
-        reason:
-            'Peace event participants = {gp1, gp2} regardless of '
-            'fromFactionId/toFactionId direction; the cooldown is '
-            'symmetric so arm 3 must still be suppressed on turn 50. '
-            'Refs #2847 § H2.',
-      );
-    });
-
-    test('Refs #2847 § H2: peace event with a different peer GP does NOT '
-        'suppress arm 3 against the blocker', () {
-      // Refs #2847 § H2 cross-peer rejection pin.
-      //
-      // The active player gp1 recently peaced gp3 (a different GP).
-      // The blocker on the GP-only frontier is gp2, with whom no
-      // peace event exists. The cooldown predicate must not trigger
-      // and arm 3 must fire on gp2.
-      final owProvinces = <Province>[
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp1_$i', regionId: 'oldWorld', ownerId: _gp1),
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp2_$i', regionId: 'oldWorld', ownerId: _gp2),
-      ];
-      final game = buildExpandGame(
-        gameIdLabel: 'expand-phase-planner-declare-war',
-        turnNumber: 50,
-        oldWorldProvinces: owProvinces,
-        armies: [
-          homeArmyWithRegimentsAtCapital(_gp1, 5),
-          homeArmyWithRegimentsAtCapital(_gp2, 5),
-        ],
-        diplomaticHistoryEvents: const [
-          DiplomaticEvent(
-            turn: 49,
-            intraTurnIndex: 0,
-            type: DiplomaticEventType.peace,
-            participants: {_gp1, _gp3},
-            fromFactionId: _gp1,
-            toFactionId: _gp3,
-          ),
-        ],
-      );
-      final snapshot = buildExpandSnapshot(
-        atWarWith: const [],
-        invadableOw: const ['oldWorld|gp2_0'],
-        adjacentOwners: const [_gp2],
-        oldWorldProvincesOwned: 8,
-      );
-      expect(
-        planExpandDeclareWar(game: game, snapshot: snapshot),
-        _gp2,
-        reason:
-            'Peace event participants = {gp1, gp3}; the blocker is gp2. '
-            'Cooldown predicate filters on participants containing the '
-            'queried peer -> rejects -> arm 3 fires as normal. Refs '
-            '#2847 § H2 cross-peer rejection.',
-      );
-    });
-
-    test('Refs #2847 S7-D H1 refutation: at-war minors that own no invadable '
-        '-> arm 2 has no candidates -> null (proximate cause is geographic '
-        'peer-war lock, not arm-2 candidate filtering)', () {
-      // Refs #2847 § S7-T H1 refutation.
-      //
-      // Pin for the seed-42 turn-99 snapshot recorded in the S7-D diagnostic
-      // for gp3 / gp4 / gp5 / gp6:
-      //
-      //   * gp3 (FAIL: +1 OW): atWarWith=[gp4, minor1, minor2, minor3, minor4,
-      //     minor6], adjacentOwnerFactionIdsSorted=[gp4], invadable=6 (all
-      //     owned by gp4).
-      //   * gp4 (FAIL: +2 OW): symmetric mirror (adjacent=[gp3], invadable
-      //     all owned by gp3).
-      //   * gp5 (FAIL: +1 OW): atWarWith=[minor1, minor2, minor4, minor6],
-      //     adjacent=[gp6], invadable owned by gp6.
-      //   * gp6 (FAIL: +2 OW): atWarWith=[minor2], adjacent=[gp5], invadable
-      //     owned by gp5.
-      //
-      // Five of the six at-war minors for gp3 (minor1..minor6) own zero
-      // invadable provinces because none of them is adjacent to gp3's
-      // anchor provinces at turn 99 — gp4's territory geographically
-      // surrounds gp3. Therefore `atWarMinors` is empty inside
-      // `planExpandDeclareWar` and arm 2 cannot fire even though
-      // `ThreatSummary.atWarWith` lists five minors.
-      //
-      // The S7-D diagnostic note ranked arm-2 candidate filtering (the
-      // `adjacentOwners` cross-check or the `atWarMinors` set construction)
-      // as H1. This pin demonstrates that the filtering is **correct**:
-      // when an at-war minor is not adjacent, `invadableProvinceIdsSorted`
-      // structurally excludes its provinces and the planner returns null
-      // by design (the spec gives no arm for at-war minors that own no
-      // invadable; conquest army-move must reach them through other
-      // territory first).
-      //
-      // S7-T scope for #2847 therefore shifts away from H1 toward the
-      // peer-war geographic lock (gp3↔gp4 and gp5↔gp6 mutually surround
-      // each other and starve each other of treasury and reachable
-      // targets). See the issue's S7-D follow-up comment and the updated
-      // S7-T tuning surface in
-      // `seed42_observer_conquest_s7d_diagnostic_test.dart`.
-      final owProvinces = <Province>[
-        for (var i = 0; i < 8; i++)
-          Province(id: 'oldWorld|gp1_$i', regionId: 'oldWorld', ownerId: _gp1),
-        for (var i = 0; i < 6; i++)
-          Province(id: 'oldWorld|gp2_$i', regionId: 'oldWorld', ownerId: _gp2),
-        Province(id: 'oldWorld|m1_a', regionId: 'oldWorld', ownerId: _minor1),
-        Province(id: 'oldWorld|m2_a', regionId: 'oldWorld', ownerId: _minor2),
-        Province(id: 'oldWorld|m3_a', regionId: 'oldWorld', ownerId: _minor3),
-      ];
-      final game = buildExpandGame(
-        gameIdLabel: 'expand-phase-planner-declare-war',
-        players: const [
-          Player(id: _gp1, displayName: 'GP1', isHuman: false, treasury: 50),
-          Player(id: _gp2, displayName: 'GP2', isHuman: false, treasury: 0),
-          Player(id: _gp3, displayName: 'GP3', isHuman: false, treasury: 9999),
-        ],
-        oldWorldProvinces: owProvinces,
-        minorNations: const [
-          MinorNation(id: _minor1, displayName: 'M1'),
-          MinorNation(id: _minor2, displayName: 'M2'),
-          MinorNation(id: _minor3, displayName: 'M3'),
-        ],
-        armies: [
-          homeArmyWithRegimentsAtCapital(_gp1, 2),
-          homeArmyWithRegimentsAtCapital(_gp2, 2),
-        ],
-      );
-      final snapshot = buildExpandSnapshot(
-        // gp1 plays the gp3 role (failing GP): at war with the peer
-        // GP2 (= gp4) **and** with several minors (m1/m2/m3) that own
-        // OW provinces but whose tiles are NOT in the invadable set
-        // because they are not P-P neighbors of gp1's anchor provinces.
-        atWarWith: const [_gp2, _minor1, _minor2, _minor3],
-        // Six invadable provinces, every one owned by the at-war peer
-        // GP2. Mirrors gp3's turn-99 snapshot.
-        invadableOw: const [
-          'oldWorld|gp2_0',
-          'oldWorld|gp2_1',
-          'oldWorld|gp2_2',
-          'oldWorld|gp2_3',
-          'oldWorld|gp2_4',
-          'oldWorld|gp2_5',
-        ],
-        // Only the at-war peer GP is geographically adjacent — the
-        // at-war minors are *not* in adjacentOwnerFactionIdsSorted, so
-        // their non-invadable OW provinces never make it into the
-        // `atWarMinors` set inside `planExpandDeclareWar`.
-        adjacentOwners: const [_gp2],
-        oldWorldProvincesOwned: 8,
-      );
-      expect(
-        planExpandDeclareWar(game: game, snapshot: snapshot),
-        isNull,
-        reason:
-            'H1 refutation pin: with all invadable OW provinces owned by '
-            'the at-war peer GP and zero at-war minors holding any of '
-            'them, arm 2 has structurally no candidates regardless of '
-            'how many minors are in ThreatSummary.atWarWith. Arm 1 is '
-            'also empty (no adjacent non-at-war minor) and arm 3 is '
-            'blocked because the sole GP frontier blocker is already '
-            'at war. The planner correctly returns null; tuning arm-2 '
-            'candidate filtering will not move the seed-42 needle. '
-            'Refs #2847 S7-T scope shift away from H1.',
-      );
-    });
   });
 }
